Add encryption and routes file

+++ scripts/deploy.sh
@@ -1,4 +1,5 @@
 user=$1
+dry_run=$2
 dist_dir=../dist
 api_dir=/var/www/api
 
@@ -7,7 +8,11 @@ mkdir -p $dist_dir
 
 cp install.sh $dist_dir/install.sh
 cp ../server.ts $dist_dir/server.ts
+cp -r ../src $dist_dir/src
 
+if [ "$dry_run" = true ]; then
+    exit
+fi
 ssh $user@smirky-games.com "rm -rf $api_dir"
 scp -r $dist_dir $user@smirky-games.com:$api_dir
 ssh $user@smirky-games.com "$api_dir/install.sh"
